@@ -1,18 +1,9 @@
-<template>
-    <div class="home">
-        <img alt="Vue logo" src="../assets/logo.png" />
-        <HelloWorld msg="Welcome to Your Vue.js + TypeScript App" />
-    </div>
-</template>
-
-<script lang="ts">
-import { Component, Vue } from "vue-property-decorator";
-import HelloWorld from "@/components/HelloWorld.vue"; // @ is an alias to /src
-
-@Component({
-    components: {
-        HelloWorld
-    }
-})
-export default class HomeView extends Vue {}
+<script setup lang="ts">
+import TheWelcome from "../components/TheWelcome.vue";
 </script>
+
+<template>
+    <main>
+        <TheWelcome />
+    </main>
+</template>
